refactor: create search module
