+++ app/modules/search/index.ts
@@ -1,0 +1,5 @@
+export { searchDrinks, purgeSearchCache } from "./implementation/search.server";
+export { SearchForm } from "./ui/search-form";
+export { NoSearchTerm } from "./ui/no-search-term";
+export { NoDrinksFound } from "./ui/no-drinks-found";
+export { Searching } from "./ui/searching";
